Fixed c++20 compilation errors and warnings
- Fixed capture warnings
- Fixed agg type errors
- Made fmt library references explicit

+++ src/giv-widget/agg/agg_renderer_outline_aa.h
@@ -1305,7 +1305,7 @@ namespace agg
             for(i = 0; i < aa_scale; i++)
             {
                 m_gamma[i] = value_type(
-                    uround(gamma_function(double(i) / aa_mask) * aa_mask));
+                  uround(gamma_function(double(i) / unsigned(aa_mask)) * unsigned(aa_mask)));
             }
         }
