+++ src/giv-widget/agg/agg_span_gradient.h
@@ -64,23 +64,23 @@ namespace agg
             m_interpolator(&inter),
             m_gradient_function(&gradient_function),
             m_color_function(&color_function),
-            m_d1(iround(d1 * gradient_subpixel_scale)),
-            m_d2(iround(d2 * gradient_subpixel_scale))
+            m_d1(iround(d1 * unsigned(gradient_subpixel_scale))),
+            m_d2(iround(d2 * unsigned(gradient_subpixel_scale)))
         {}
 
         //--------------------------------------------------------------------
         interpolator_type& interpolator() { return *m_interpolator; }
         const GradientF& gradient_function() const { return *m_gradient_function; }
         const ColorF& color_function() const { return *m_color_function; }
-        double d1() const { return double(m_d1) / gradient_subpixel_scale; }
-        double d2() const { return double(m_d2) / gradient_subpixel_scale; }
+        double d1() const { return double(m_d1) / unsigned(gradient_subpixel_scale); }
+        double d2() const { return double(m_d2) / unsigned(gradient_subpixel_scale); }
 
         //--------------------------------------------------------------------
         void interpolator(interpolator_type& i) { m_interpolator = &i; }
         void gradient_function(const GradientF& gf) { m_gradient_function = &gf; }
         void color_function(const ColorF& cf) { m_color_function = &cf; }
-        void d1(double v) { m_d1 = iround(v * gradient_subpixel_scale); }
-        void d2(double v) { m_d2 = iround(v * gradient_subpixel_scale); }
+        void d1(double v) { m_d1 = iround(v * unsigned(gradient_subpixel_scale)); }
+        void d2(double v) { m_d2 = iround(v * unsigned(gradient_subpixel_scale)); }
 
         //--------------------------------------------------------------------
         void prepare() {}
@@ -197,9 +197,9 @@ namespace agg
 
         //---------------------------------------------------------------------
         gradient_radial_focus(double r, double fx, double fy) : 
-            m_r (iround(r  * gradient_subpixel_scale)), 
-            m_fx(iround(fx * gradient_subpixel_scale)), 
-            m_fy(iround(fy * gradient_subpixel_scale))
+            m_r (iround(r  * unsigned(gradient_subpixel_scale))), 
+            m_fx(iround(fx * unsigned(gradient_subpixel_scale))), 
+            m_fy(iround(fy * unsigned(gradient_subpixel_scale)))
         {
             update_values();
         }
@@ -207,16 +207,16 @@ namespace agg
         //---------------------------------------------------------------------
         void init(double r, double fx, double fy)
         {
-            m_r  = iround(r  * gradient_subpixel_scale);
-            m_fx = iround(fx * gradient_subpixel_scale);
-            m_fy = iround(fy * gradient_subpixel_scale);
+            m_r  = iround(r  * unsigned(gradient_subpixel_scale));
+            m_fx = iround(fx * unsigned(gradient_subpixel_scale));
+            m_fy = iround(fy * unsigned(gradient_subpixel_scale));
             update_values();
         }
 
         //---------------------------------------------------------------------
-        double radius()  const { return double(m_r)  / gradient_subpixel_scale; }
-        double focus_x() const { return double(m_fx) / gradient_subpixel_scale; }
-        double focus_y() const { return double(m_fy) / gradient_subpixel_scale; }
+        double radius()  const { return double(m_r)  / unsigned(gradient_subpixel_scale); }
+        double focus_x() const { return double(m_fx) / unsigned(gradient_subpixel_scale); }
+        double focus_y() const { return double(m_fy) / unsigned(gradient_subpixel_scale); }
 
         //---------------------------------------------------------------------
         int calculate(int x, int y, int) const
